Move new topology graphs from example to its own module.
