+++ docs/source/topologies.rst
@@ -1,0 +1,14 @@
+Topologies
+==========
+
+A package of the classes for handling topology graphs in `CGExplore`.
+
+New topology graphs
+-------------------
+
+.. toctree::
+  :maxdepth: 1
+
+  CGM12L24 <_autosummary/cgexplore.topologies.CGM12L24>
+  CGM4L8 <_autosummary/cgexplore.topologies.CGM4L8>
+  M4L82 <_autosummary/cgexplore.topologies.M4L82>
