+++ src/cgexplore/topologies.py
@@ -1,0 +1,9 @@
+"""molecular package."""
+
+from cgexplore._internal.topologies.graphs import CGM4L8, CGM12L24, M4L82
+
+__all__ = [
+    "CGM12L24",
+    "CGM4L8",
+    "M4L82",
+]
